third_party/WebKit/LayoutTests/fast/dom/Document/document-charset.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=koi8-r">
</head>
<body>
<script>

if (window.testRunner)
    testRunner.dumpAsText();

document.write("<p>Initial</p>");
document.write("<p>&nbsp;document.charset: " + document.charset + "</p>"); // MSIE
document.write("<p>&nbsp;document.characterSet: " + document.characterSet + "</p>"); // Firefox
document.write("<p>&nbsp;document.inputEncoding: " + document.inputEncoding + "</p>"); // DOM3

document.write("<p>Setting charset to UTF-8... (expected to have no effect)</p>");
document.charset = "utf-8";
document.write("<p>&nbsp;document.charset: " + document.charset + "</p>");
document.write("<p>&nbsp;document.characterSet: " + document.characterSet + "</p>");
document.write("<p>&nbsp;document.inputEncoding: " + document.inputEncoding + "</p>");

</script>
</body>
<html>
